Remove usage of uint32_t

+++ src/Attribute_Request_Callback.cpp
@@ -23,11 +23,11 @@ Attribute_Request_Callback::Attribute_Request_Callback(const char *attributes, f
 
 #endif // !THINGSBOARD_ENABLE_STL
 
-const uint32_t& Attribute_Request_Callback::Get_Request_ID() const {
+const size_t& Attribute_Request_Callback::Get_Request_ID() const {
     return m_request_id;
 }
 
-void Attribute_Request_Callback::Set_Request_ID(const uint32_t &request_id) {
+void Attribute_Request_Callback::Set_Request_ID(const size_t &request_id) {
     m_request_id = request_id;
 }
